Add a default derive
This is somewhat hard to use because there's no way to know whether the
default should apply to the current item or not. Currently, to use the
old behavior, you need to explicitly say `merge(strategy =
Merge::merge)`. Otherwise, it will try to use the default and cause a
compile error:

```rust
error[E0308]: mismatched types
   --> tests/derive.rs:552:21
    |
552 |     #[derive(Debug, Merge, PartialEq)]
    |                     ^^^^^ expected enum `std::option::Option`, found struct `test_default_strategy::N`
    |
    = note: expected mutable reference `&mut std::option::Option<_>`
               found mutable reference `&mut test_default_strategy::N`
    = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0308]: mismatched types
   --> tests/derive.rs:552:21
    |
552 |     #[derive(Debug, Merge, PartialEq)]
    |                     ^^^^^
    |                     |
    |                     expected enum `std::option::Option`, found struct `test_default_strategy::N`
    |                     help: try using a variant of the expected enum: `Some(Merge)`
    |
    = note: expected enum `std::option::Option<_>`
             found struct `test_default_strategy::N`
    = note: this error originates in a derive macro (in Nightly builds, run with -Z macro-backtrace for more info)
```

I'm open to suggestions on how to make this easier to use/give better
error messages.